--- src/main/kotlin/app/revanced/patches/music/utils/fix/client/SpoofSignaturePatch.kt
@@ -1,5 +1,0 @@
-package app.revanced.patches.music.utils.fix.client
-
-import app.revanced.patches.shared.spoofsignature.BaseSpoofSignaturePatch
-
-object SpoofSignaturePatch : BaseSpoofSignaturePatch()
