+++ src/main/kotlin/app/revanced/patches/music/utils/gms/GmsCoreSupportPatch.kt
@@ -1,7 +1,6 @@
 package app.revanced.patches.music.utils.gms
 
 import app.revanced.patches.music.utils.compatibility.Constants.COMPATIBLE_PACKAGE
-import app.revanced.patches.music.utils.fix.client.SpoofSignaturePatch
 import app.revanced.patches.music.utils.fix.client.SpoofUserAgentPatch
 import app.revanced.patches.music.utils.fix.fileprovider.FileProviderPatch
 import app.revanced.patches.music.utils.integrations.IntegrationsPatch
@@ -16,7 +15,6 @@ object GmsCoreSupportPatch : BaseGmsCoreSupportPatch(
     mainActivityOnCreateFingerprint = MainActivityFingerprint,
     integrationsPatchDependency = IntegrationsPatch::class,
     dependencies = setOf(
-        SpoofSignaturePatch::class,
         SpoofUserAgentPatch::class,
         PackageNamePatch::class,
         FileProviderPatch::class
