+++ src/main/kotlin/app/revanced/patches/shared/gms/fingerprints/GmsCoreSupportFingerprint.kt
@@ -3,9 +3,8 @@ package app.revanced.patches.shared.gms.fingerprints
 import app.revanced.patcher.fingerprint.MethodFingerprint
 
 internal object GmsCoreSupportFingerprint : MethodFingerprint(
-    customFingerprint = { _, classDef ->
-        classDef.type.endsWith("GmsCoreSupport;")
+    customFingerprint = { methodDef, _ ->
+        methodDef.definingClass.endsWith("/GmsCoreSupport;")
+                && methodDef.name == "getGmsCoreVendorGroupId"
     },
-) {
-    const val GET_GMS_CORE_VENDOR_GROUP_ID_METHOD_NAME = "getGmsCoreVendorGroupId"
-}
+)
