+++ src/main/resources/youtube/settings/xml/revanced_prefs.xml
@@ -611,9 +611,6 @@
         <SwitchPreference android:title="@string/revanced_spoof_player_parameter_title" android:key="revanced_spoof_player_parameter" android:summary="@string/revanced_spoof_player_parameter_summary" />
         <SwitchPreference android:title="@string/revanced_spoof_player_parameter_in_feed_title" android:key="revanced_spoof_player_parameter_in_feed" android:summaryOn="@string/revanced_spoof_player_parameter_in_feed_summary_on" android:summaryOff="@string/revanced_spoof_player_parameter_in_feed_summary_off" android:dependency="revanced_spoof_player_parameter" />SETTINGS: SPOOF_PLAYER_PARAMETER -->
 
-        <!-- SETTINGS: SPOOF_SIGNATURE
-        <SwitchPreference android:title="@string/revanced_spoof_signature_title" android:key="revanced_spoof_signature" android:summaryOn="@string/revanced_spoof_signature_summary_on" android:summaryOff="@string/revanced_spoof_signature_summary_off" />SETTINGS: SPOOF_SIGNATURE -->
-
         <PreferenceCategory android:title="@string/revanced_preference_category_patch_information" android:layout="@layout/revanced_settings_preferences_category"/>
         <PreferenceScreen android:title="@string/revanced_preference_screen_patch_information_title"  android:key="revanced_preference_screen_patch_information" android:summary="@string/revanced_preference_screen_patch_information_summary" >
             <PreferenceCategory android:title="@string/revanced_preference_category_tool_used" android:layout="@layout/revanced_settings_preferences_category" >
